backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
apscheduler==3.10.4
python-dotenv==1.0.1
python-multipart==0.0.12
httpx>=0.27.0
twilio>=9.0.0
